requirements.txt
ansicon==1.89.0
appdirs==1.4.4
atomicwrites==1.4.0
attrs==20.3.0
blessed==1.18.0
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
Flask==1.1.2
Flask-WTF==0.14.3
flyingcircus==0.1.3.2
gast==0.4.0
google-auth==1.28.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
gunicorn==20.1.0
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
jinxed==1.0.1
joblib==1.0.1
Markdown==3.3.4
MarkupSafe==1.1.1
mysql==0.0.2
mysql-connector-python-rf==2.2.2
mysqlclient==2.0.3
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.9
pathlib==1.0.1
pluggy==0.13.1
protobuf==3.15.7
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.3
python-dotenv==0.17.0
PyYAML==5.4.1
regex==2021.3.17
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
setuptools-scm==6.0.1
six==1.15.0
toml==0.10.2
urllib3==1.26.2
wcwidth==0.2.5
Werkzeug==1.0.1
WTForms==2.3.3
